DDBTEAM-461: Fixed event handling for post/delete material
